Add vendored System.Diagnostics.DiagnosticSource (#30)
This brings a vendored version of System.Diagnostics.DiagnosticSource to the repo. This can be further optimized to make fewer allocations but it is operational.

To easily compare the changes vs dotnet runtime compare the folders against dotnet/runtime at:

661ac50e084 Fix the closure allocations in ActivityTagsCollection (#42606)
Compare the following folders:

<OTel.NET.Instr-repo>\src\OpenTelemetry.DynamicActivityBinding\vendored\System\Diagnostics\
<dotnet\runtime-repo>\src\libraries\System.Diagnostics.DiagnosticSource\src\System\Diagnostics\
Main changes:

Span to byte[]
Span to char[] and string this requires changing calls from AsSpan to ToString and Substring as appropriate.
To avoid Unsafe.AsRef some structs and fields had "readonly" removed.