Small build improvements (#1531)
* Increase dependency warmup timeout to reduce flakiness

* Ignore CMake temporary build files (macos)

* Don't need the static.a file in the tracer home

* Simplify/Unify the build tracer home step for linux + mac

Previously we were copying files _into_ the native project's bin folder, and then publishing from there. This change makes things a bit simpler in the pipeline as well as being more consistent wih the output in general.

- Only publish the net4 targets in the managed build step on Windows (not required for linux)
- Copy the files _to_ the tracer home directly,  _from_ the bin output, instead of the other way around

* Add build step for generating the tracer home directory in the dd-trace format

dd-trace uses a different folder layout for dd-trace.
This adds an extra Nuke target for creating a directory with the required structure (required for Linux and Mac)

* Add pipeline variable for ddTracerHome

* Fix comment

* Try fix flakiness in Arm64 Linux Integration builds

* Suppress build warnings on macos