Merge tools into Nuke (#1605)
* Merge GeneratePackageVersions into Nuke

* Move UpdateVendors tool to be a Nuke target

Embed the template as a resource for simplicity

* minor: move Nuke helper files into folder for easier navigation

* Move EnvironmentTools.cs to Datadog.Trace.TestHelpers

It's only used in test projects anyway

* Move TcpPortProvider to TestHelpers, as again, it's only used in tests

Some of the sample projects use this type, so rather than reference TestHelpers (and hence transitively Datadog.Trace), added the helper class a link file instead.

* Remove remaining references to Datadog.Core.Tools

* Move PrepareRelease to Nuke

This required quite a lot of work in the GenerateIntegrationDefinitions code to use reflection instead of direct references.

Additionally, I purposefully iterate all the built versions of the assembly, and de-dupe the duplicates, to avoid cases where some integrations are not available on some platforms. This required a slightly annoying amount of boilerplate and faffing with AppLoadContexts

* Move CIAppTracking project to Nuke

* Don't override everything in Nuke's editorconfig

* Update build/_build/Build.Utilities.cs

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>

* Remove DistributedTracer tool

It's for a very specific use-case, and as it's the last remaining tool, I decided to remove it

* Apply suggestions from code review

Co-authored-by: Lucas Pimentel-Ordyna <lucas.pimentel@datadoghq.com>

* Update some target names

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>
Co-authored-by: Lucas Pimentel-Ordyna <lucas.pimentel@datadoghq.com>