Add GitHub workflows for automating Release creation (#1611)
This creates 4 new workflows:
* Auto add PR to `vNext` milestone
* Create version bump PR
* Auto tag version bump commit
* Create draft release from latest tag

## Auto add PR to vNext milestone

To make later stages of the release workflow easier, this will automatically run when a pull request is closed, and add it to a `vNext` milestone. Only _merged_ PRs that will be included in the next release should be added to the `vNext` milestone.

## Create version bump PR

When we're ready to do a release, the first step is to do a "version bump" commit. This workflow can be _manually_ triggered from the GitHub actions page, where you provide the next version number, and whether or not it's a prerelease.

The action first updates the changelog, using the titles of all PRs included in the `vNext` milestone, using the existing format.

Next it does the version bump, updates integrations.json, and syncs the MSI contents.

After that's complete it renames the `vNext` milestone to the appropriate value, e.g. `1.28.2-prerelease`.

Finally, it creates a branch called `version-bump-1.28.2-prerelease` (for example), commits the changed files to it, pushes, and opens a PR.

Users can then review the PR, make changes if required, and merge the PR.

## Auto tag version bump commit

When the `[Version Bump]` commit above is merged to master, a new workflow is kicked off, that automatically pushes a new tag, `v1.28.2-prerelease` to the repo. This also creates the benchmark branch (as the previous worklfow was doing)

 ## Create draft release from latest tag

 The final workflow creates the draft release from the latest tag assets. This assumes that you've already done a version bump PR (it doesn't _require_ that it's the auto-bump, it just needs to have an entry in the changelog, and assets in Azure Devops).

 The workflow starts by downloading the assets for the `v1.28.2-prerelease` tag build from Azure Devops. We publish a "well known" artifact that it uses to pull everything down, warning/erroring if it can't find an expected build, if the build isn't complete, or if it failed to produce any assets.

 Once download and extracted, it extracts the release notes from the changelog, and uses these to create a draft release. The release notes include warning/reminders about uploading the nuget packages to nuget, and about attaching the signed MSI artifacts

 Final publishing is up to a user, there's no automation around that