Add fix for log injection with log4net.Ext.Json versions < 2.0.9.1 (#1615)
Add fix for log4net.Ext.Json versions < 2.0.9.1 not supporting log4net.Core.IFixingRequired for special-case object rendering.

Background
The log4net.Ext.Json library is a prominent third-party library that can render log4net objects into JSON. To prepare the final JSON object, the StandardTypesDecorator does a type check against each value in the MDC dictionary and tries to render each object accordingly. If the value is a well-known type, then the idiomatic JSON value is written. If not, the decorator will render the value as a new JSON object and attempt to render each of the original object's members as properties on the new JSON object.

Bug
In log4net.Ext.Json versions < 2.0.9.1, the interface log4net.Core.IFixingRequired is not included in the set of types specifically recognized as a built-in type, so the custom object Log4NetEnricher+TracerProperty that we insert into the MDC dictionary is rendered as a new JSON object, not a string, which breaks trace-log correlation.

Fix
Before using our CustomLog4NetLogProvider (which utilizes our Log4NetEnricher for holding the current trace_id and span_id values for logs injection), check if the application is using log4net.Ext.Json version 2.0.9.1 or newer. If not, we'll fallback to the original liblog log provider implementation.