Add better error handling for the Header Tags feature accessing System.Web.HttpResponse.Headers (#1641)
Fixes "System.PlatformNotSupportedException: This operation requires IIS integrated pipeline mode"

This scenario can occur even when the IIS application is running in the Integrated pipeline mode, which seems to be related to Sitefinity CMS widgets. Without having a reproduction available, this doesn't prevent the bad access from happening. Instead, we have better handling of the scenario:
1. Only attempt to access the System.Web.HttpResponse.Headers field when the application has explicitly configured the DD_TRACE_HEADER_TAGS/TracerSettings.HeaderTags feature, which avoids the issue in some cases
2. Catch the exception and prevents further field accesses when the exception is observed