The v0 version of App Sec (#1647)
This PR implements a first version of App Sec. Please forgive the message history, it will be squashed away :)

The App Sec module introduces two new environment variables to activate App Sec features:

`DD_APPSEC_ENABLED` – enables the App Sec feature, any HTTP requests will be examined by the WAF (Web Application Firewall). By default, this feature is disabled.

`DD_APPSEC_BLOCKING_ENABLED` – this controls whether the App Sec feature will block suspicious requests. By default, this feature is disabled.

The app sec uses the ‘call target’ instrumentation to gather the data about the HTTP request that will be passed to the WAF. It is has not been implemented for ‘call site’ instrumentation and currently there is not plan to support this style of instrumentation.

Data is gathered in the instrumentations for ASP.NET and ASP.NET Core and then passed to the ‘instrumentation gateway’ which starts the security flow. Currently the instrumentation gateway simply creates and event with the incoming data. In future version this component will be used to route the data to different security components based on the addresses associated with the data, but this functionality is not necessary for the first version.

After the instrumentation gateway the data flow is as follows:

```
InstrumentationGateway -> Security -> PowerWaf -> Native -> Sqreen.dll
```

`Security` – this class coordinates the reaction to security events
`PowerWaf` – this class provides .NET idiomatic wrapper to the WAF, as the WAF is implemented in C++ with C style interface
`Native` – this class is responsible for finding, loading and providing low level bindings with the WAF
`[lib]Sqreen.[dll|so|dylib]` – this are the native binaries containing the WAF implementation.

Security events are reported to the agent and backend by `AgentWriter` class, which implements a basic queuing system.

Two new samples are provided, that form the basis for the security integration tests. These are `Samples.AspNetMvc5` and `Samples.AspNetCore5` which are used to test ASP.NET and ASP.NET Core respectively.

Integration tests based on similar principals to `Datadog.Trace.ClrProfiler.IntegrationTests` have been provided in `Datadog.Trace.Security.IntegrationTests` and are run as part of the integration tests in the `consolidated-pipeline` CI.

ARM based Linux systems are not supported yet, but all other environment supported by the .NET libraries are.

ASP.NET Core running on .NET Framework is not supported. The instrumentation is a no-op in these scenarios. This is because it though this unlikely to work, since we rely on referencing several `Microsoft.AspNetCore.*` dlls to be able instrument the ASP.NET Core pipeline.

Co-authored-by: Anna Yafi <anna.yafi@datadoghq.com>
Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>