Merging repos: 'dd-shared-components-dotnet' into 'dd-trace-dotnet'. (#1694)
This is the first of several self-contained PRs targeted at moving the Shared Code (as in _shared between the Tracer and the Profiler_) from [dd-shared-components-dotnet](https://github.com/DataDog/dd-shared-components-dotnet) into [dd-trace-dotnet](https://github.com/DataDog/dd-trace-dotnet).

For this PR, I created a branch `permanent/MoveFrom_dd-shared-components-dotnet` and moved all files from the `dd-shared-components-dotnet` into there, including their complete change history. The branch will remain permanently as a history record, and we will squash-merge this PR into the `master` branch. The files are all inside the `/shared` folder, so that they can be naturally merged into the planned joint directory structure.

In addition, this PR also makes the following small changes:

* Make whatever tweaks necessary so that _all_ the Shared Code projects build inside `dd-trace-dotnet`.
* Set up placeholder folders for the `/tracer` and the `/profiler` products to facilitate the directory refactoring in the subsequent PRs.
* Update ReadMe files to describe the changes.

Note that this PR keeps the directory structure within the `/shared` folder exactly as it was within the original repo. Subsequent changed will adjust this structure to joint conventions as required. This will require coordination with the profiler build configuration which makes some assumptions about the layout of the relevant folders.

This change is self-contained. It will not affect any existing Tracer components, and it will allow the overall repo merge project to proceed in several steps that can be validated independently.

This PR will be followed up with additional PR to address the following:

* Move Tracer-specific files into the `/tracer` folder.
* Ensure that changes contained to shared files that are not (yet) used by the Tracer do not require the Tracer CI validation to pass (this will be updated in the future if and when the Tracer uses more shared components)
* Joint build of components used by several products.
* Joint release package.
* Joint conventions for folder structure.
* Etc...