Minor tweaks to shared files (#1729)
As we are cleaning up the Profiler build to reference `dd-trace-dotnet` instead of `dd-shared-components-dotnet`, we discover the need to minor tweaks. This PR is the collection of such tweaks which I collected into a single PR to minimize churn.

The Profiler work is now done (https://github.com/DataDog/dd-continuous-profiler-dotnet/pull/135).
Note that while this PR is being reviewed and merged, the Profiler builds cleanly while referencing `dd-trace-dotnet / macrogreg/SharedItemsCatchup`. Once this PR is merged, we must point the profiler to `dd-trace-dotnet / master`.

Changes here:

* Make packages path the same for all shared items.
* Props files must use relative paths to construct shared file references.
* Fix upward walking chains in `Directory.Build.props` files.